Inheritance of MapSearchService -> TextSearchService -> SearchableService.
