+++ samples/Wave.Searchability/src/Wave.Searchability/packages.config
@@ -3,5 +3,5 @@
   <package id="Extended.Wpf.Toolkit" version="2.5" targetFramework="net35" />
   <package id="log4net" version="1.2.10" targetFramework="net35" />
   <package id="TaskParallelLibrary" version="1.0.2856.0" targetFramework="net35" />
-  <package id="Wave.9.3" version="0.1.50-alpha" targetFramework="net35" />
+  <package id="Wave.9.3" version="0.1.51-alpha" targetFramework="net35" />
 </packages>
